chore: Update build script to copy files to public folder
This commit updates the build script in the `package.json` file to include a command that copies the files from the `dist` directory to the `public` directory. This ensures that the built files are accessible to the application when it is deployed.

It also includes changes to the service worker file to add event listeners for the `install`, `activate`, and `fetch` events. These event listeners handle the installation and activation of the service worker, as well as the caching and fetching of resources for whitelisted hostnames.

In addition, the `webShareAPI.tsx` file has been modified to import the `useBrowserConditions` hook from the correct location.

Lastly, the `index.html` file has been updated to include additional Apple touch icons and meta tags for improved mobile support and social media sharing.

This commit improves the build process and enhances the functionality and performance of the application.